Working on Monseuir LeDsp...

+++ interfaces/mce_library/context.h
@@ -68,8 +68,8 @@ typedef struct mcecmd {
     int (*connect)(mce_context_t);
     int (*disconnect)(mce_context_t);
     int (*ioctl)(mce_context_t, unsigned long int, int);
-    ssize_t (*read)(mce_context_t, void*, size_t);
-    ssize_t (*write)(mce_context_t, const void*, size_t);
+    int (*read)(mce_context_t, void*, size_t);
+    int (*write)(mce_context_t, const void*, size_t);
 } mcecmd_t;
 
 typedef enum {
@@ -96,8 +96,8 @@ struct mcedsp_context {
     int (*connect)(mce_context_t);
     int (*disconnect)(mce_context_t);
     int (*ioctl)(mce_context_t, unsigned long int, ...);
-    ssize_t (*read)(mce_context_t, void*, size_t);
-    ssize_t (*write)(mce_context_t, const void*, size_t);
+    int (*read)(mce_context_t, void*, size_t);
+    int (*write)(mce_context_t, const void*, size_t);
 };
 typedef struct mcedsp_context mcedsp_t;
