+++ interfaces/mce_library/net.h
@@ -13,8 +13,8 @@ int mcenet_hello(mce_context_t context);
 extern int mcecmd_net_connect(mce_context_t);
 extern int mcecmd_net_disconnect(mce_context_t);
 extern int mcecmd_net_ioctl(mce_context_t, unsigned long int, int);
-extern ssize_t mcecmd_net_read(mce_context_t, void*, size_t);
-extern ssize_t mcecmd_net_write(mce_context_t, const void*, size_t);
+extern int mcecmd_net_read(mce_context_t, void*, size_t);
+extern int mcecmd_net_write(mce_context_t, const void*, size_t);
 
 extern int mcedata_net_connect(mce_context_t);
 extern int mcedata_net_disconnect(mce_context_t);
@@ -25,7 +25,7 @@ extern ssize_t mcedata_net_write(mce_context_t, const void*, size_t);
 extern int mcedsp_net_connect(mce_context_t);
 extern int mcedsp_net_disconnect(mce_context_t);
 extern int mcedsp_net_ioctl(mce_context_t, unsigned long int, ...);
-extern ssize_t mcedsp_net_read(mce_context_t, void*, size_t);
-extern ssize_t mcedsp_net_write(mce_context_t, const void*, size_t);
+extern int mcedsp_net_read(mce_context_t, void*, size_t);
+extern int mcedsp_net_write(mce_context_t, const void*, size_t);
 
 #endif
