+++ interfaces/mce_library/sdsu.h
@@ -10,8 +10,8 @@
 extern int mcecmd_sdsu_connect(mce_context_t);
 extern int mcecmd_sdsu_disconnect(mce_context_t);
 extern int mcecmd_sdsu_ioctl(mce_context_t, unsigned long int, int);
-extern ssize_t mcecmd_sdsu_read(mce_context_t, void*, size_t);
-extern ssize_t mcecmd_sdsu_write(mce_context_t, const void*, size_t);
+extern int mcecmd_sdsu_read(mce_context_t, void*, size_t);
+extern int mcecmd_sdsu_write(mce_context_t, const void*, size_t);
 
 extern int mcedata_sdsu_connect(mce_context_t);
 extern int mcedata_sdsu_disconnect(mce_context_t);
@@ -22,7 +22,7 @@ extern ssize_t mcedata_sdsu_write(mce_context_t, const void*, size_t);
 extern int mcedsp_sdsu_connect(mce_context_t);
 extern int mcedsp_sdsu_disconnect(mce_context_t);
 extern int mcedsp_sdsu_ioctl(mce_context_t, unsigned long int, ...);
-extern ssize_t mcedsp_sdsu_read(mce_context_t, void*, size_t);
-extern ssize_t mcedsp_sdsu_write(mce_context_t, const void*, size_t);
+extern int mcedsp_sdsu_read(mce_context_t, void*, size_t);
+extern int mcedsp_sdsu_write(mce_context_t, const void*, size_t);
 
 #endif
